[accumulators] Move a lot of code from accumulator.hpp to .cpp files
This would be impossible without removing support for user-defined
observable types (ALPS_ACCUMULATOR_USER_TYPES macro and custom_type
test).

This commit effectively reverts solution to #157.
